added unit tests and some housekeeping

+++ server/services/ProductService.js
@@ -5,15 +5,16 @@ let Product = require("../models/product");
 
 
 // CHALLENGE 1
-recordRoutes.route("/api/products").post(function (req, res) {
+recordRoutes.post("/api/products",function (req, res) {
     var db = dbo.getDB();
+    if (!req.body.id) throw "Please input an Id";
+
     let paramid = {_id: req.body.id};
     delete req.body.id;
     req.body.createdAt = new Date(req.body.createdAt);
     req.body.updatedAt = new Date(req.body.updatedAt);
     let param = Object.assign(paramid, req.body);
 
-    console.log(param)
     db.collection("Products").insertOne(param, function (err, apiResponse) {
       if (err) throw err;
       var result = new Product(Object.assign({id: paramid._id}, req.body));
@@ -25,6 +26,8 @@ recordRoutes.route("/api/products").post(function (req, res) {
 // CHALLENGE 2
 recordRoutes.route("/api/products/:id").get(function (req, res) {
     var db = dbo.getDB();
+    if (!req.params.id) throw "Please input an Id";
+
     let param = { _id: req.params.id };
     db.collection("Products").findOne(param).then(function(result) {
       result = new Product(Object.assign({id: param._id},result));
@@ -35,6 +38,9 @@ recordRoutes.route("/api/products/:id").get(function (req, res) {
 // CHALLENGE 3
 recordRoutes.route("/api/products/:id").post(function (req, res) {
     var db = dbo.getDB();
+    if (!req.body.id || !req.body.createdAt || !req.body.updatedAt || req.body.name 
+      || req.body.price) throw "Please input Id, createdAt, updatedAt, name and price";
+
     let newvalues = {
       $set: {
         "createdAt": new Date(req.body.createdAt),
@@ -55,9 +61,12 @@ recordRoutes.route("/api/products/:id").post(function (req, res) {
     });
 
   });
+  
 // CHALLENGE 4
 recordRoutes.route("/api/products/:id").delete((req, res) => {
     var db = dbo.getDB();
+    if (!req.params.id) throw "Please input an Id";
+
     var param = { _id: req.params.id };
     db.collection("Products").deleteOne(param, function (err, obj) {
       if (err) throw err;
